% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_construction.R
\name{gforce.certify}
\alias{gforce.certify}
\title{K-means optimality certificate.}
\usage{
gforce.certify(sol, D, eps1 = 0.01, eps2 = 10^-7, Y_T_min = 0.01)
}
\arguments{
\item{sol}{vector of length \eqn{d}. This contains the assignment of variables or
points to clusters.}

\item{D}{\eqn{d x d} matrix.}

\item{eps1}{a scalar. It controls the stopping condition for the dual solution search.}

\item{eps2}{a scalar. It controls the infeasibility tolerance for the dual solution to allow for numerical imprecision.}

\item{Y_T_min}{a scalar. THe smallest \eqn{Y_T} that the function can return. Must be greater than 0.}
}
\description{
Given a proposed integer solution to the \eqn{K}-means SDP relaxation, this
function attempts to construct a solution to the dual problem with matching
objective value.
}

