% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust2mat.R
\name{gforce.clust2mat}
\alias{gforce.clust2mat}
\title{Convert K-means solution to partnership matrix.}
\usage{
gforce.clust2mat(clusters)
}
\arguments{
\item{clusters}{length \eqn{d} vector. Assigns each variable or data point to a cluster. Cluster names
can be numbers or strings.}
}
\description{
Convert K-means solution to partnership matrix.
}

