% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_control.R
\name{gforce.confint2test}
\alias{gforce.confint2test}
\title{Convert confidence intervals to equivalent test statistics.}
\usage{
gforce.confint2test(conf_ints, alpha)
}
\arguments{
\item{conf_ints}{symmetric matrix of confidence intervals.}

\item{alpha}{alpha level of the confidence intervals.}
}
\description{
Can convert a 4D array encoding the confidence intervals for a precision
matrix to standard normal test-statistics.
}

