% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORCE.R
\name{gforce.defaults}
\alias{gforce.defaults}
\title{Provides the default tuning parameters for \code{\link{gforce.FORCE}}.}
\usage{
gforce.defaults(d, K)
}
\arguments{
\item{d}{dimension of random vector or number of datapoints.}

\item{K}{number of clusters.}
}
\description{
Provides the default tuning parameters for \code{\link{gforce.FORCE}}.
}

