% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORCE.init.R
\name{gforce.FORCE.init}
\alias{gforce.FORCE.init}
\title{Generate Random Initial Iterate and Values for FORCE Algorithm.}
\usage{
gforce.FORCE.init(diff, K, s, opt_estimate, R_only = TRUE)
}
\arguments{
\item{K}{number of groups.}

\item{s}{numeric value. Indicates the initial mixing of \code{opt_estimate} and a random full rank solution.}

\item{opt_estimate}{\eqn{d x d} numeric array. Specifies a guess of a feasible solution and is used to construct the initial feasible solution.}

\item{D}{\eqn{d x d} numeric array. Specifies the objective function of the \eqn{K}-means SDP.}
}
\value{
An object with following components
\describe{
\item{\code{E}}{Strictly feasible solution.}
\item{\code{X0}}{Strictly feasible initial iterate.}
\item{\code{E_val}}{Objective value of SDP at \eqn{E}.}
\item{\code{X0_val}}{Objective value of SDP at \eqn{X_0}.}
}
}
\description{
This function constructs a random full-rank solution to the Peng-Wei \eqn{K}-means SDP
and also gives an initial iterate \eqn{X0} for the FORCE algorithm.
}
\examples{
K <- 5
n <- 50 
d <- 50
m <- 3
s <- 0.25
dat <- gforce.generator(K,d,n,m,graph='DeltaC')
sh <- t(dat$X)\%*\%dat$X / n
gh <- gforce.Gamma(dat$X,par=TRUE)
D <- diag(gh)-sh
opt_guess <- kmeans(sh,K)$cluster
init_vals <- gforce.FORCE.init(D,K,s)


}

