% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORCE.R
\name{gforce.FORCE}
\alias{gforce.FORCE}
\title{Solves the Peng-Wei K-means SDP Relaxation using the FORCE algorithm.}
\usage{
gforce.FORCE(D, K, force_opts = NULL, D_Kmeans = NULL, X0 = NULL,
  E = NULL)
}
\arguments{
\item{D}{a matrix \eqn{D} as defined above.}

\item{K}{number of clusters.}

\item{force_opts}{tuning parameters. \code{NULL} signifies defaults will be used.}

\item{D_Kmeans}{matrix to be used for initial integer solution. \code{NULL} signifies that \code{D} will be used.}

\item{X0}{initial iterate. \code{NULL} signifies that it will be generated randomly from \code{D_Kmeans}. If supplied, \code{E} must be supplied as well.}

\item{E}{strictly feasible solutions. \code{NULL} signifies that it will be generated randomly. If supplied, \code{X0} must be supplied as well.}
}
\description{
Solves the Peng-Wei K-means SDP Relaxation using the FORCE algorithm.
}

