% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORCE.init.R
\name{gforce.FORCE_adapt.init}
\alias{gforce.FORCE_adapt.init}
\title{Generate Random Initial Iterate and Values for FORCE Algorithm with Adaptive Selection of \eqn{K}.}
\usage{
gforce.FORCE_adapt.init(D, K, s, opt_estimate, R_only = FALSE,
  cluster_representation = FALSE, par = FALSE)
}
\arguments{
\item{D}{\eqn{d x d} numeric array. Specifies the objective function of the \eqn{K}-means SDP.}

\item{K}{number of groups.}

\item{s}{numeric value. Indicates the initial mixing of \code{opt_estimate} and a random full rank solution.}

\item{opt_estimate}{\eqn{d x d} numeric array. Specifies a guess of a feasible solution and is used to construct the initial feasible solution.}

\item{R_only}{logical expression. If \code{R_only == FALSE}, then the included
native code implementation will be used. Otherwise, an R implementation is used.}

\item{cluster_representation}{logical expression. If \code{cluster_representation == FALSE}, then \code{opt_estimate} is assumed to
be a length \eqn{d} array with the values \eqn{0,...,K-1}, indicating cluster assignments.}

\item{par}{logical expression. If \code{par == TRUE}, then a multi-threaded version
of the function is called. If \code{par == FALSE}, a single-threaded version is called.}
}
\value{
An object with following components
\describe{
\item{\code{X0}}{Strictly feasible initial iterate.}
\item{\code{X0_val}}{Objective value of SDP at \eqn{X_0}.}
}
}
\description{
This function constructs a random full-rank solution to the Peng-Wei \eqn{K}-means SDP
and also gives an initial iterate \eqn{X0} for the FORCE algorithm.
}
\examples{
K <- 5
n <- 50 
d <- 50
m <- 3
s <- 0.25
dat <- gforce.generator(K,d,n,m,graph='DeltaC')
sh <- t(dat$X)\%*\%dat$X / n
gh <- gforce.Gamma(dat$X,par=TRUE)
D <- diag(gh)-sh
opt_guess <- kmeans(sh,K)$cluster
init_vals <- gforce.FORCE_adapt.init(D,K,s)


}

