% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_sdp_matrix.R
\name{gforce.FORCE_matrix_rep}
\alias{gforce.FORCE_matrix_rep}
\title{FORCE SDP Matrices.}
\usage{
gforce.FORCE_matrix_rep(D, K)
}
\arguments{
\item{K}{number of clusters.}

\item{d}{dimension.}
}
\value{
An object with following components
\describe{
\item{\code{Z_T}}{Final iterate of the projected gradient descent algorithm run on the smoothed eigenvalue problem.}
\item{\code{B_Z_T}}{Projection of \code{Z_T} to the border of the positive semi-definite cone.}
}
}
\description{
Constructs the SDP constraint matrices so they can be passed to a black-box solver.
}
\examples{
K <- 5
d <- 50

}

