% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_rank_feasible.R
\name{gforce.full_rank_feasible}
\alias{gforce.full_rank_feasible}
\title{Strictly Feasible Solution to \eqn{K}-means SDP.}
\usage{
gforce.full_rank_feasible(d, K, R_only = FALSE)
}
\arguments{
\item{d}{dimension of the solution.}

\item{K}{number of groups.}

\item{R_only}{logical expression. If \code{R_only == TRUE}, then no native code is run.}
}
\value{
A strictly feasible solution \eqn{E}.
}
\description{
This function constructs a  full-rank solution to the Peng-Wei \eqn{K}-means SDP.
}
\examples{
K <- 5
n <- 50 
d <- 50
E <- gforce.full_rank_feasible(d,K)

}

