% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PECOK.R
\name{gforce.Gamma}
\alias{gforce.Gamma}
\title{Estimates \eqn{\Gamma} for the PECOK SDP.}
\usage{
gforce.Gamma(X, par = FALSE)
}
\arguments{
\item{X}{\eqn{n x d} matrix. Each row represents a realization of a
\eqn{d} dimensional random vector.}

\item{par}{logical expression. If \code{par == TRUE}, then a multi-threaded version
of the function is called. If \code{par == FALSE}, a single-threaded version is called.}
}
\description{
In particular, it returns in diagonal form the estimator
\eqn{\Gamma} used to construct the PECOK penalized covariance estimator.
}

