% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{gforce.generator}
\alias{gforce.generator}
\title{Data generator.}
\usage{
gforce.generator(K, d, n, m, graph = "DeltaC", num_hubs = NULL,
  cov_gap_mult = 1, error_base = 0.25, error_add = 0.25,
  corr_value = 0.3)
}
\arguments{
\item{K}{number of clusters.}

\item{d}{dimension of the observed random vector.}

\item{n}{number of samples.}

\item{m}{minimal group size.}

\item{graph}{latent graph structure. Can be 'scalefree', 'hub', 'band' or 'DeltaC'.}

\item{num_hubs}{number of hubs in the latent graph. Ignored unless \code{graph == 'hub'}.}

\item{cov_gap_mult}{scales the size of \eqn{\delta C}. Ignored unless \code{graph == 'DeltaC'}.}

\item{error_base}{minimum variance of \eqn{E_i}.}

\item{error_add}{size of range of possible variances for \eqn{E_i}.}

\item{corr_value}{size of off diagonal entries in latent precision matrix.}
}
\description{
Generates \eqn{n} random samples from a \eqn{G}-Latent Variable Model. The
caller can specify the graph structure on the latent variables via
several parameters. The magnitude of the non-zero entries in the population
precision matrix can also be specified. Observed variables are assigned
uniformly at random to \eqn{K} groups with minimum size \eqn{m}.
}

