% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatent_confints.R
\name{gforce.glatent_confints.cv_defaults}
\alias{gforce.glatent_confints.cv_defaults}
\title{Default Cross Validation Options for Confidence Intervals.}
\usage{
gforce.glatent_confints.cv_defaults()
}
\description{
Default Cross Validation Options for Confidence Intervals.
}

