% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glatent_confints.R
\name{gforce.glatent_confints}
\alias{gforce.glatent_confints}
\title{Confidence Intervals for Estimation in G-Latent Models.}
\usage{
gforce.glatent_confints(C_hat = NULL, X_vals = NULL, clusters = NULL,
  alpha = 0.05, graph = "latent", use_cv = FALSE, cv_opts = NULL,
  lambda1 = NULL, lambda2 = NULL, use_scio_package = TRUE)
}
\arguments{
\item{K}{number of clusters.}
}
\description{
Estimate the precision matrix and construct confidence intervals for the latent and group averages graphs. The user
can either provide an estimate of the relevant covariance matrix or the data and cluster assignments. If cross validation
is selected, the data and clusters must be provided.
}

