% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{gforce.hclust.agg2clust}
\alias{gforce.hclust.agg2clust}
\title{Hierarchical Clustering with Estimation of \eqn{K}.}
\usage{
gforce.hclust.agg2clust(hc, K)
}
\description{
Hierarchical Clustering with Estimation of \eqn{K}.
}

