% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{gforce.hclust.agglomerate}
\alias{gforce.hclust.agglomerate}
\title{Hierarchical Clustering with Estimation of \eqn{K}.}
\usage{
gforce.hclust.agglomerate(X = NULL, dists = NULL)
}
\description{
Hierarchical Clustering with Estimation of \eqn{K}.
}

