% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{gforce.hclustCagg}
\alias{gforce.hclustCagg}
\title{Hierarchical Clustering with Estimation of \eqn{K}.}
\usage{
gforce.hclustCagg(hc, K)
}
\description{
Hierarchical Clustering with Estimation of \eqn{K}.
}

