% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{gforce.kmeans}
\alias{gforce.kmeans}
\title{K-means Clustering.}
\usage{
gforce.kmeans(X, K, R_only = FALSE)
}
\arguments{
\item{X}{\eqn{n x m} matrix. Each row is treated as a point in \eqn{R^m}.}

\item{K}{integer. The number of clusters to group the data into.}

\item{R_only}{logical expression. If \code{R_only == FALSE}, then the included
native code implementation will be used. Otherwise, an R implementation is used.}
}
\description{
Solves the K-means problem using kmeans++ for the initialization and then 
runs Lloyd's algorithm.
}

