% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FORCE.R
\name{gforce.PECOK}
\alias{gforce.PECOK}
\title{Uses the FORCE algorithm to solve the PECOK SDP.}
\usage{
gforce.PECOK(K, X = NULL, D = NULL, force_opts = NULL, X0 = NULL,
  E = NULL)
}
\arguments{
\item{K}{number of clusters.}

\item{X}{\eqn{n x d} matrix. Either this or \code{D} must be specified.}

\item{D}{\eqn{d x d} matrix. Either this or \code{X} must be specified.}

\item{force_opts}{tuning parameters. \code{NULL} signifies defaults will be used.}

\item{X0}{initial iterate. \code{NULL} signifies that it will be generated randomly from \code{D_Kmeans}. If supplied, \code{E} must be supplied as well.}

\item{E}{strictly feasible solutions. \code{NULL} signifies that it will be generated randomly. If supplied, \code{X0} must be supplied as well.}
}
\description{
Uses the FORCE algorithm to solve the PECOK SDP.
}
\seealso{
\code{\link{gforce.FORCE.options}}
}

