% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scio.R
\name{gforce.scio}
\alias{gforce.scio}
\title{SCIO Estimator.}
\usage{
gforce.scio(C, lambda, k = NULL, eps = 10^-6, max_iter = 10000,
  R_only = FALSE)
}
\arguments{
\item{K}{number of clusters.}
}
\description{
Estimate the precision matrix with the SCIO estimator. 
The implementation follows the active set strategy also used in the SCIO package.
}

